19- New modal added to db

+++ app/api/clients/[clientId]/route.ts
@@ -27,14 +27,49 @@ export async function PATCH(
   }
 }
 
+// export async function DELETE(
+//   request: Request,
+//   { params }: { params: { clientId: string } }
+// ) {
+//   try {
+//     const { userId } = await auth();
+//     const { clientId } = await params;
+//     if (!userId) return new NextResponse("Unauthorized", { status: 401 });
+
+//     if (!clientId) {
+//       return NextResponse.json(
+//         { error: "Client ID is required!" },
+//         { status: 400 }
+//       );
+//     }
+
+//     await db.client.delete({
+//       where: { id: clientId },
+//     });
+
+//     return NextResponse.json(
+//       { message: "Client deleted successfully" },
+//       { status: 200 }
+//     );
+//   } catch (error) {
+//     return NextResponse.json(
+//       { error: "Failed to delete client" },
+//       { status: 500 }
+//     );
+//   }
+// }
+
 export async function DELETE(
   request: Request,
   { params }: { params: { clientId: string } }
 ) {
   try {
     const { userId } = await auth();
-    const { clientId } = await params;
-    if (!userId) return new NextResponse("Unauthorized", { status: 401 });
+    const { clientId } = params;
+
+    if (!userId) {
+      return new NextResponse("Unauthorized", { status: 401 });
+    }
 
     if (!clientId) {
       return NextResponse.json(
@@ -43,6 +78,22 @@ export async function DELETE(
       );
     }
 
+    // Check if the client has unpaid purchases
+    const unpaidPurchases = await db.purchase.findMany({
+      where: {
+        clientId,
+        paymentStatus: "UNPAID",
+      },
+    });
+
+    if (unpaidPurchases.length > 0) {
+      return NextResponse.json(
+        { error: "Can't delete client with unpaid purchases." },
+        { status: 400 }
+      );
+    }
+
+    // Proceed with deleting the client
     await db.client.delete({
       where: { id: clientId },
     });
@@ -52,6 +103,7 @@ export async function DELETE(
       { status: 200 }
     );
   } catch (error) {
+    console.error("Error deleting client:", error);
     return NextResponse.json(
       { error: "Failed to delete client" },
       { status: 500 }
